+++ src/helpers/error-handler.js
@@ -1,4 +1,5 @@
 function errorHanlder(err, req, res, next) {
+  console.log("err", err);
   if (err.name === "UnauthorizedError") {
     return res.status(401).json({ message: "User is not Authorized" });
   }
