+++ src/routers/order.js
@@ -1,5 +1,12 @@
-const { Order } = require("../models/order");
-const { OrderItem } = require("../models/order-item");
+const {
+  listOrder,
+  orderDetail,
+  createOrder,
+  updateOrder,
+  delelteOrder,
+  totalSales,
+  getOrderByUser,
+} = require("../controller/order");
 const express = require("express");
 const router = express.Router();
 
@@ -66,7 +73,7 @@ const router = express.Router();
 
 /**
  * @swagger
- * /order:
+ * /api/v1/order:
  *   get:
  *     summary: Returns the list of all the order
  *     tags: [Order]
@@ -81,20 +88,11 @@ const router = express.Router();
  *
  */
 
-router.get("/", async (req, res) => {
-  const orderList = await Order.find()
-    .populate("user", "name")
-    .sort({ dateOrdered: -1 });
-
-  if (!orderList) {
-    res.status(500).send("Cannot get order");
-  }
-  res.send(orderList);
-});
+router.get("/", listOrder);
 
 /**
  * @swagger
- * /order/{id}:
+ * /api/v1/order/{id}:
  *   get:
  *     summary: Get the order by id
  *     tags: [Order]
@@ -115,29 +113,11 @@ router.get("/", async (req, res) => {
  *         description: Cannot get order detail
  */
 
-router.get("/:id", async (req, res) => {
-  const productDetail = await Order.findById(req.params.id)
-    .populate("user", "name")
-    .populate({
-      path: "orderItems",
-      populate: {
-        path: "product",
-        populate: {
-          path: "category",
-        },
-      },
-    });
-
-  if (!productDetail) {
-    res.status(500).send("Cannot get order detail");
-  }
-
-  res.send(productDetail);
-});
+router.get("/:id", orderDetail);
 
 /**
  * @swagger
- * /order:
+ * /api/v1/order:
  *   post:
  *     summary: create new order
  *     tags: [Order]
@@ -158,57 +138,11 @@ router.get("/:id", async (req, res) => {
  *         description: Some server error
  */
 
-router.post("/", async (req, res) => {
-  const orderItemIds = Promise.all(
-    req.body.orderItems.map(async (orderItem) => {
-      let newOrderItem = new OrderItem({
-        quantity: orderItem.quantity,
-        product: orderItem.product,
-      });
-      newOrderItem = await newOrderItem.save();
-      return newOrderItem._id;
-    })
-  );
-
-  const orderItemList = await orderItemIds;
-
-  const totalPrices = await Promise.all(
-    orderItemList.map(async (orderItemId) => {
-      const orderItem = await OrderItem.findById(orderItemId).populate(
-        "product",
-        "price"
-      );
-      const totalPrice = orderItem.product.price * orderItem.quantity;
-      return totalPrice;
-    })
-  );
-
-  const totalPrice = totalPrices.reduce((a, b) => a + b, 0);
-
-  let order = new Order({
-    orderItems: orderItemList,
-    shippingAddress1: req.body.shippingAddress1,
-    shippingAddress2: req.body.shippingAddress2,
-    city: req.body.city,
-    zip: req.body.zip,
-    country: req.body.country,
-    phone: req.body.phone,
-    status: req.body.status,
-    totalPrice: totalPrice,
-    user: req.body.user,
-  });
-
-  order = await order.save();
-
-  if (!order) {
-    return res.status(400).send("Cannot create order !");
-  }
-  res.send(order);
-});
+router.post("/", createOrder);
 
 /**
  *  @swagger
- * /order/{id}:
+ * /api/v1/order/{id}:
  *  put:
  *    summary: Update the order by the id
  *    tags: [Order]
@@ -238,25 +172,11 @@ router.post("/", async (req, res) => {
  *        description: Some error happened
  */
 
-router.put("/:id", async (req, res) => {
-  const orderUpdate = await Order.findByIdAndUpdate(
-    req.params.id,
-    {
-      status: req.body.status,
-    },
-    { new: true }
-  );
-
-  if (!orderUpdate) {
-    res.status(500).send("Order cannot update");
-  }
-
-  res.send(orderUpdate);
-});
+router.put("/:id", updateOrder);
 
 /**
  * @swagger
- * /order/{id}:
+ * /api/v1/order/{id}:
  *   delete:
  *     summary: Remove the order by id
  *     tags: [Order]
@@ -275,28 +195,11 @@ router.put("/:id", async (req, res) => {
  *         description: The order was not found
  */
 
-router.delete("/:id", (req, res) => {
-  Order.findByIdAndRemove(req.params.id)
-    .then(async (order) => {
-      if (order) {
-        await order.orderItems.map(async (orderitemId) => {
-          await OrderItem.findByIdAndRemove(orderitemId);
-        });
-        return res
-          .status(200)
-          .send({ success: true, message: "Order deleted successfully" });
-      } else {
-        return res.status(401).send("Invalid Order");
-      }
-    })
-    .catch((err) => {
-      return res.status(500).send({ success: false, error: err });
-    });
-});
+router.delete("/:id", delelteOrder);
 
 /**
  * @swagger
- * /order/get/totalSale:
+ * /api/v1/order/get/totalSale:
  *   get:
  *     summary: Get total sale
  *     tags: [Order]
@@ -312,21 +215,11 @@ router.delete("/:id", (req, res) => {
  *                 $ref: '#/components/schemas/Order'
  */
 
-router.get("/get/totalSale", async (req, res) => {
-  const totalSales = await Order.aggregate([
-    { $group: { _id: null, totalSales: { $sum: "$totalPrice" } } },
-  ]);
-
-  if (!totalSales) {
-    return res.status(500).send("Cannot get total of sale");
-  }
-  console.log("totalSales", totalSales);
-  res.send({ totalSales: totalSales.pop().totalSales });
-});
+router.get("/get/totalSale", totalSales);
 
 /**
  * @swagger
- * /order/get/orderbyuser/{userId}:
+ * /api/v1/order/get/orderbyuser/{userId}:
  *   get:
  *     summary: Get list order by userid
  *     tags: [Order]
@@ -348,26 +241,6 @@ router.get("/get/totalSale", async (req, res) => {
  *                 $ref: '#/components/schemas/Order'
  */
 
-router.get("/get/orderbyuser/:userId", async (req, res) => {
-  const orderListByUser = await Order.find({
-    user: req.params.userId,
-  })
-    .populate("user", "name")
-    .populate({
-      path: "orderItems",
-      populate: {
-        path: "product",
-        populate: {
-          path: "category",
-        },
-      },
-    })
-    .sort({ dateOrdered: -1 });
-
-  if (!orderListByUser) {
-    res.status(500).send("Cannot get order");
-  }
-  res.send(orderListByUser);
-});
+router.get("/get/orderbyuser/:userId", getOrderByUser);
 
 module.exports = router;
