+++ src/routers/products.js
@@ -1,9 +1,15 @@
-const { Category } = require("../models/category");
-const { Product } = require("../models/products");
 const express = require("express");
 const router = express.Router();
-const mongoose = require("mongoose");
-const { handleUpload, getFileName } = require("../helpers/upload");
+const {
+  productList,
+  productDetail,
+
+  createNewproduct,
+  updateProduct,
+  deleteProduct,
+  productCount,
+  getFeatures,
+} = require("../controller/product");
 
 /**
  * @swagger
@@ -79,18 +85,7 @@ const { handleUpload, getFileName } = require("../helpers/upload");
  *
  */
 
-router.get("/", async (req, res) => {
-  // filter product by multiple categories
-  let filter = {};
-  if (req.query.categories) {
-    filter = { category: req.query.categories.split(",") };
-  }
-  const productList = await Product.find(filter).populate("category");
-  if (!productList) {
-    res.status(500).json({ success: false });
-  }
-  res.send(productList);
-});
+router.get("/", productList);
 
 /**
  * @swagger
@@ -115,17 +110,7 @@ router.get("/", async (req, res) => {
  *         description: Cannot get product detail
  */
 
-router.get("/:id", async (req, res) => {
-  const productDetail = await Product.findById(req.params.id).populate(
-    "category"
-  );
-
-  if (!productDetail) {
-    res.status(500).json({ success: false });
-  }
-
-  res.send(productDetail);
-});
+router.get("/:id", productDetail);
 
 /**
  * @swagger
@@ -150,33 +135,7 @@ router.get("/:id", async (req, res) => {
  *         description: Some server error
  */
 
-router.post(`/`, async (req, res) => {
-  const category = await Category.findById(req.body.category);
-  if (!category) return res.status(400).send("Invalid Category");
-
-  // const imagePath = await getFileName(req.body.image, res);
-  // const image = await handleUpload(imagePath);
-
-  let product = new Product({
-    name: req.body.name,
-    description: req.body.description,
-    richDescription: req.body.richDescription,
-    image: req.body.image,
-    brand: req.body.brand,
-    price: req.body.price,
-    category: req.body.category,
-    countInStock: req.body.countInStock,
-    rating: req.body.rating,
-    numReviews: req.body.numReviews,
-    isFeatured: req.body.isFeatured,
-  });
-
-  product = await product.save();
-
-  if (!product) return res.status(500).send("The product cannot be created");
-
-  res.send(product);
-});
+router.post(`/`, createNewproduct);
 
 /**
  *  @swagger
@@ -210,37 +169,7 @@ router.post(`/`, async (req, res) => {
  *        description: Some error happened
  */
 
-router.put(`/:id`, async (req, res) => {
-  if (!mongoose.isValidObjectId(req.params.id)) {
-    return res.status(400).send("Invalid product");
-  }
-  const category = await Category.findById(req.body.category);
-  if (!category) return res.status(400).send("Invalid Category");
-
-  const product = await Product.findByIdAndUpdate(
-    req.params.id,
-    {
-      name: req.body.name,
-      description: req.body.description,
-      richDescription: req.body.richDescription,
-      image: req.body.image,
-      brand: req.body.brand,
-      price: req.body.price,
-      category: req.body.category,
-      countInStock: req.body.countInStock,
-      rating: req.body.rating,
-      numReviews: req.body.numReviews,
-      isFeatured: req.body.isFeatured,
-    },
-    {
-      new: true,
-    }
-  );
-
-  if (!product) return res.status(500).send("The product cannot be updated");
-
-  res.send(product);
-});
+router.put(`/:id`, updateProduct);
 
 /**
  * @swagger
@@ -263,23 +192,7 @@ router.put(`/:id`, async (req, res) => {
  *         description: The product was not found
  */
 
-router.delete("/:id", (req, res) => {
-  Product.findByIdAndRemove(req.params.id)
-    .then((product) => {
-      if (product) {
-        return res
-          .status(200)
-          .json({ success: true, message: "The product is deleted!" });
-      } else {
-        return res
-          .status(404)
-          .json({ success: false, message: "Product not found!" });
-      }
-    })
-    .catch((err) => {
-      return res.status(500).json({ success: false, error: err });
-    });
-});
+router.delete("/:id", deleteProduct);
 
 /**
  * @swagger
@@ -299,16 +212,7 @@ router.delete("/:id", (req, res) => {
  *                 $ref: '#/components/schemas/Product'
  */
 
-router.get("/get/count", async (req, res) => {
-  const productCount = await Product.countDocuments();
-
-  if (!productCount) {
-    res.status(500).json({ success: false });
-  }
-  res.send({
-    productCount: productCount,
-  });
-});
+router.get("/get/count", productCount);
 
 /**
  * @swagger
@@ -334,14 +238,6 @@ router.get("/get/count", async (req, res) => {
  *                 $ref: '#/components/schemas/Product'
  */
 
-router.get("/get/featured/:count", async (req, res) => {
-  const count = req.params.count ? req.params.count : 0;
-  const products = await Product.find({ isFeatured: true }).limit(+count);
-
-  if (!products) {
-    res.status(500).json({ success: false });
-  }
-  res.send(products);
-});
+router.get("/get/featured/:count", getFeatures);
 
 module.exports = router;
